- Moving MCU initializations file
- init peripheral file was created
