+++ host_comm_fsm/.settings/stm32cubeide.project.prefs
@@ -1,4 +1,5 @@
 635E684B79701B039C64EA45C3F84D30=2AF0D7A53E30A7D805C16883916A3A13
 66BE74F758C12D739921AEA421D593D3=1
-DC22A860405A8BF2F2C095E5B6529F12=5BCAA8B579D4DCF1B85005ADFA1F6293
+8DF89ED150041C4CBC7CB9A9CAA90856=BEBF1E3E724C30096EEF51D3640D59F2
+DC22A860405A8BF2F2C095E5B6529F12=E8CD77FF6F7F74E28A8F4466F12B5F17
 eclipse.preferences.version=1
