+++ host_comm_fsm/.vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "files.associations": {
+        "init_peripherals.h": "c"
+    }
+}
