+++ host_comm_fsm/Core/Src/stm32f4xx_hal_msp.c
@@ -18,7 +18,7 @@
   */
 
 /* Includes ------------------------------------------------------------------*/
-#include "main.h"
+#include "init_peripherals.h"
 
 
 /* Private typedef -----------------------------------------------------------*/
