+++ host_comm_fsm/Debug/Core/Src/init_peripherals.su
@@ -1,0 +1,4 @@
+init_peripherals.c:12:6:SystemClock_Config	88	static
+init_peripherals.c:57:13:MX_USART2_UART_Init	8	static
+init_peripherals.c:78:13:MX_GPIO_Init	48	static
+init_peripherals.c:110:6:peripherals_init	8	static
