+++ host_comm_fsm/Debug/Core/Src/main.su
@@ -1,0 +1,2 @@
+main.c:20:5:main	8	static
+main.c:36:6:Error_Handler	4	static,ignoring_inline_asm
