+++ host_comm_fsm/Debug/Core/Src/stm32f4xx_hal_msp.su
@@ -1,0 +1,3 @@
+stm32f4xx_hal_msp.c:36:6:HAL_MspInit	16	static
+stm32f4xx_hal_msp.c:52:6:HAL_UART_MspInit	48	static
+stm32f4xx_hal_msp.c:85:6:HAL_UART_MspDeInit	16	static
