+++ host_comm_fsm/Debug/Core/Src/stm32f4xx_it.su
@@ -1,0 +1,10 @@
+stm32f4xx_it.c:29:6:NMI_Handler	4	static
+stm32f4xx_it.c:39:6:HardFault_Handler	4	static
+stm32f4xx_it.c:49:6:MemManage_Handler	4	static
+stm32f4xx_it.c:59:6:BusFault_Handler	4	static
+stm32f4xx_it.c:69:6:UsageFault_Handler	4	static
+stm32f4xx_it.c:79:6:SVC_Handler	4	static
+stm32f4xx_it.c:87:6:DebugMon_Handler	4	static
+stm32f4xx_it.c:95:6:PendSV_Handler	4	static
+stm32f4xx_it.c:103:6:SysTick_Handler	8	static
+stm32f4xx_it.c:118:6:USART2_IRQHandler	8	static
