+++ host_comm_fsm/Debug/Core/Src/syscalls.su
@@ -1,0 +1,18 @@
+syscalls.c:45:6:initialise_monitor_handles	4	static
+syscalls.c:49:5:_getpid	4	static
+syscalls.c:54:5:_kill	16	static
+syscalls.c:60:6:_exit	16	static
+syscalls.c:66:27:_read	32	static
+syscalls.c:78:27:_write	32	static
+syscalls.c:89:5:_close	16	static
+syscalls.c:95:5:_fstat	16	static
+syscalls.c:101:5:_isatty	16	static
+syscalls.c:106:5:_lseek	24	static
+syscalls.c:111:5:_open	12	static
+syscalls.c:117:5:_wait	16	static
+syscalls.c:123:5:_unlink	16	static
+syscalls.c:129:5:_times	16	static
+syscalls.c:134:5:_stat	16	static
+syscalls.c:140:5:_link	16	static
+syscalls.c:146:5:_fork	8	static
+syscalls.c:152:5:_execve	24	static
