+++ host_comm_fsm/Debug/Core/Src/sysmem.su
@@ -1,0 +1,1 @@
+sysmem.c:54:7:_sbrk	32	static
