+++ host_comm_fsm/Debug/Core/Src/system_stm32f4xx.su
@@ -1,0 +1,2 @@
+system_stm32f4xx.c:168:6:SystemInit	4	static
+system_stm32f4xx.c:221:6:SystemCoreClockUpdate	32	static
