+++ host_comm_fsm/Debug/Drivers/STM32F4xx_HAL_Driver/Src/stm32f4xx_hal.su
@@ -1,0 +1,27 @@
+stm32f4xx_hal.c:157:19:HAL_Init	8	static
+stm32f4xx_hal.c:190:19:HAL_DeInit	8	static
+stm32f4xx_hal.c:219:13:HAL_MspInit	4	static
+stm32f4xx_hal.c:230:13:HAL_MspDeInit	4	static
+stm32f4xx_hal.c:253:26:HAL_InitTick	16	static
+stm32f4xx_hal.c:312:13:HAL_IncTick	4	static
+stm32f4xx_hal.c:323:17:HAL_GetTick	4	static
+stm32f4xx_hal.c:332:10:HAL_GetTickPrio	4	static
+stm32f4xx_hal.c:341:19:HAL_SetTickFreq	24	static
+stm32f4xx_hal.c:373:21:HAL_GetTickFreq	4	static
+stm32f4xx_hal.c:389:13:HAL_Delay	24	static
+stm32f4xx_hal.c:415:13:HAL_SuspendTick	4	static
+stm32f4xx_hal.c:431:13:HAL_ResumeTick	4	static
+stm32f4xx_hal.c:441:10:HAL_GetHalVersion	4	static
+stm32f4xx_hal.c:450:10:HAL_GetREVID	4	static
+stm32f4xx_hal.c:459:10:HAL_GetDEVID	4	static
+stm32f4xx_hal.c:468:6:HAL_DBGMCU_EnableDBGSleepMode	4	static
+stm32f4xx_hal.c:477:6:HAL_DBGMCU_DisableDBGSleepMode	4	static
+stm32f4xx_hal.c:486:6:HAL_DBGMCU_EnableDBGStopMode	4	static
+stm32f4xx_hal.c:495:6:HAL_DBGMCU_DisableDBGStopMode	4	static
+stm32f4xx_hal.c:504:6:HAL_DBGMCU_EnableDBGStandbyMode	4	static
+stm32f4xx_hal.c:513:6:HAL_DBGMCU_DisableDBGStandbyMode	4	static
+stm32f4xx_hal.c:524:6:HAL_EnableCompensationCell	4	static
+stm32f4xx_hal.c:535:6:HAL_DisableCompensationCell	4	static
+stm32f4xx_hal.c:544:10:HAL_GetUIDw0	4	static
+stm32f4xx_hal.c:553:10:HAL_GetUIDw1	4	static
+stm32f4xx_hal.c:562:10:HAL_GetUIDw2	4	static
