+++ host_comm_fsm/Debug/Drivers/STM32F4xx_HAL_Driver/Src/stm32f4xx_hal_cortex.su
@@ -1,0 +1,32 @@
+core_cm4.h:1657:22:__NVIC_SetPriorityGrouping	24	static
+core_cm4.h:1676:26:__NVIC_GetPriorityGrouping	4	static
+core_cm4.h:1688:22:__NVIC_EnableIRQ	16	static
+core_cm4.h:1724:22:__NVIC_DisableIRQ	16	static,ignoring_inline_asm
+core_cm4.h:1743:26:__NVIC_GetPendingIRQ	16	static
+core_cm4.h:1762:22:__NVIC_SetPendingIRQ	16	static
+core_cm4.h:1777:22:__NVIC_ClearPendingIRQ	16	static
+core_cm4.h:1794:26:__NVIC_GetActive	16	static
+core_cm4.h:1816:22:__NVIC_SetPriority	16	static
+core_cm4.h:1838:26:__NVIC_GetPriority	16	static
+core_cm4.h:1863:26:NVIC_EncodePriority	40	static
+core_cm4.h:1890:22:NVIC_DecodePriority	40	static
+core_cm4.h:1939:34:__NVIC_SystemReset	4	static,ignoring_inline_asm
+core_cm4.h:2022:26:SysTick_Config	16	static
+stm32f4xx_hal_cortex.c:143:6:HAL_NVIC_SetPriorityGrouping	16	static
+stm32f4xx_hal_cortex.c:165:6:HAL_NVIC_SetPriority	32	static
+stm32f4xx_hal_cortex.c:187:6:HAL_NVIC_EnableIRQ	16	static
+stm32f4xx_hal_cortex.c:203:6:HAL_NVIC_DisableIRQ	16	static
+stm32f4xx_hal_cortex.c:216:6:HAL_NVIC_SystemReset	8	static
+stm32f4xx_hal_cortex.c:229:10:HAL_SYSTICK_Config	16	static
+stm32f4xx_hal_cortex.c:258:6:HAL_MPU_Disable	4	static,ignoring_inline_asm
+stm32f4xx_hal_cortex.c:281:6:HAL_MPU_Enable	16	static,ignoring_inline_asm
+stm32f4xx_hal_cortex.c:300:6:HAL_MPU_ConfigRegion	16	static
+stm32f4xx_hal_cortex.c:344:10:HAL_NVIC_GetPriorityGrouping	8	static
+stm32f4xx_hal_cortex.c:371:6:HAL_NVIC_GetPriority	24	static
+stm32f4xx_hal_cortex.c:386:6:HAL_NVIC_SetPendingIRQ	16	static
+stm32f4xx_hal_cortex.c:404:10:HAL_NVIC_GetPendingIRQ	16	static
+stm32f4xx_hal_cortex.c:420:6:HAL_NVIC_ClearPendingIRQ	16	static
+stm32f4xx_hal_cortex.c:437:10:HAL_NVIC_GetActive	16	static
+stm32f4xx_hal_cortex.c:454:6:HAL_SYSTICK_CLKSourceConfig	16	static
+stm32f4xx_hal_cortex.c:472:6:HAL_SYSTICK_IRQHandler	8	static
+stm32f4xx_hal_cortex.c:481:13:HAL_SYSTICK_Callback	4	static
