+++ host_comm_fsm/Debug/Drivers/STM32F4xx_HAL_Driver/Src/stm32f4xx_hal_exti.su
@@ -1,0 +1,9 @@
+stm32f4xx_hal_exti.c:143:19:HAL_EXTI_SetConfigLine	32	static
+stm32f4xx_hal_exti.c:238:19:HAL_EXTI_GetConfigLine	32	static
+stm32f4xx_hal_exti.c:317:19:HAL_EXTI_ClearConfigLine	32	static
+stm32f4xx_hal_exti.c:370:19:HAL_EXTI_RegisterCallback	32	static
+stm32f4xx_hal_exti.c:395:19:HAL_EXTI_GetHandle	16	static
+stm32f4xx_hal_exti.c:435:6:HAL_EXTI_IRQHandler	24	static
+stm32f4xx_hal_exti.c:467:10:HAL_EXTI_GetPending	32	static
+stm32f4xx_hal_exti.c:496:6:HAL_EXTI_ClearPending	24	static
+stm32f4xx_hal_exti.c:517:6:HAL_EXTI_GenerateSWI	24	static
