+++ host_comm_fsm/Debug/Drivers/STM32F4xx_HAL_Driver/Src/stm32f4xx_hal_flash_ex.su
@@ -1,0 +1,22 @@
+stm32f4xx_hal_flash_ex.c:162:19:HAL_FLASHEx_Erase	24	static
+stm32f4xx_hal_flash_ex.c:233:19:HAL_FLASHEx_Erase_IT	24	static
+stm32f4xx_hal_flash_ex.c:286:19:HAL_FLASHEx_OBProgram	24	static
+stm32f4xx_hal_flash_ex.c:345:6:HAL_FLASHEx_OBGetConfig	16	static
+stm32f4xx_hal_flash_ex.c:374:19:HAL_FLASHEx_AdvOBProgram	24	static
+stm32f4xx_hal_flash_ex.c:430:6:HAL_FLASHEx_AdvOBGetConfig	16	static
+stm32f4xx_hal_flash_ex.c:462:19:HAL_FLASHEx_OB_SelectPCROP	16	static
+stm32f4xx_hal_flash_ex.c:487:19:HAL_FLASHEx_OB_DeSelectPCROP	16	static
+stm32f4xx_hal_flash_ex.c:953:13:FLASH_MassErase	16	static
+stm32f4xx_hal_flash_ex.c:982:6:FLASH_Erase_Sector	24	static
+stm32f4xx_hal_flash_ex.c:1032:26:FLASH_OB_EnableWRP	24	static
+stm32f4xx_hal_flash_ex.c:1068:26:FLASH_OB_DisableWRP	24	static
+stm32f4xx_hal_flash_ex.c:1101:26:FLASH_OB_EnablePCROP	24	static
+stm32f4xx_hal_flash_ex.c:1129:26:FLASH_OB_DisablePCROP	24	static
+stm32f4xx_hal_flash_ex.c:1162:26:FLASH_OB_RDP_LevelConfig	24	static
+stm32f4xx_hal_flash_ex.c:1196:26:FLASH_OB_UserConfig	24	static
+stm32f4xx_hal_flash_ex.c:1231:26:FLASH_OB_BOR_LevelConfig	16	static
+stm32f4xx_hal_flash_ex.c:1249:16:FLASH_OB_GetUser	4	static
+stm32f4xx_hal_flash_ex.c:1259:17:FLASH_OB_GetWRP	4	static
+stm32f4xx_hal_flash_ex.c:1273:16:FLASH_OB_GetRDP	16	static
+stm32f4xx_hal_flash_ex.c:1301:16:FLASH_OB_GetBOR	4	static
+stm32f4xx_hal_flash_ex.c:1311:6:FLASH_FlushCaches	4	static
