+++ host_comm_fsm/Debug/Drivers/STM32F4xx_HAL_Driver/Src/stm32f4xx_hal_flash_ramfunc.su
@@ -1,0 +1,4 @@
+stm32f4xx_hal_flash_ramfunc.c:96:30:HAL_FLASHEx_StopFlashInterfaceClk	16	static
+stm32f4xx_hal_flash_ramfunc.c:113:30:HAL_FLASHEx_StartFlashInterfaceClk	16	static
+stm32f4xx_hal_flash_ramfunc.c:130:30:HAL_FLASHEx_EnableFlashSleepMode	16	static
+stm32f4xx_hal_flash_ramfunc.c:147:30:HAL_FLASHEx_DisableFlashSleepMode	16	static
