+++ host_comm_fsm/Debug/Drivers/STM32F4xx_HAL_Driver/Src/stm32f4xx_hal_pwr.su
@@ -1,0 +1,17 @@
+stm32f4xx_hal_pwr.c:92:6:HAL_PWR_DeInit	4	static
+stm32f4xx_hal_pwr.c:110:6:HAL_PWR_EnableBkUpAccess	16	static
+stm32f4xx_hal_pwr.c:130:6:HAL_PWR_DisableBkUpAccess	16	static
+stm32f4xx_hal_pwr.c:268:6:HAL_PWR_ConfigPVD	16	static
+stm32f4xx_hal_pwr.c:311:6:HAL_PWR_EnablePVD	4	static
+stm32f4xx_hal_pwr.c:320:6:HAL_PWR_DisablePVD	4	static
+stm32f4xx_hal_pwr.c:334:6:HAL_PWR_EnableWakeUpPin	16	static
+stm32f4xx_hal_pwr.c:352:6:HAL_PWR_DisableWakeUpPin	16	static
+stm32f4xx_hal_pwr.c:381:6:HAL_PWR_EnterSLEEPMode	16	static,ignoring_inline_asm
+stm32f4xx_hal_pwr.c:424:6:HAL_PWR_EnterSTOPMode	16	static,ignoring_inline_asm
+stm32f4xx_hal_pwr.c:463:6:HAL_PWR_EnterSTANDBYMode	4	static,ignoring_inline_asm
+stm32f4xx_hal_pwr.c:484:6:HAL_PWR_PVD_IRQHandler	8	static
+stm32f4xx_hal_pwr.c:501:13:HAL_PWR_PVDCallback	4	static
+stm32f4xx_hal_pwr.c:516:6:HAL_PWR_EnableSleepOnExit	4	static
+stm32f4xx_hal_pwr.c:528:6:HAL_PWR_DisableSleepOnExit	4	static
+stm32f4xx_hal_pwr.c:540:6:HAL_PWR_EnableSEVOnPend	4	static
+stm32f4xx_hal_pwr.c:552:6:HAL_PWR_DisableSEVOnPend	4	static
