+++ host_comm_fsm/Debug/Drivers/STM32F4xx_HAL_Driver/Src/stm32f4xx_hal_pwr_ex.su
@@ -1,0 +1,10 @@
+stm32f4xx_hal_pwr_ex.c:143:19:HAL_PWREx_EnableBkUpReg	16	static
+stm32f4xx_hal_pwr_ex.c:167:19:HAL_PWREx_DisableBkUpReg	16	static
+stm32f4xx_hal_pwr_ex.c:191:6:HAL_PWREx_EnableFlashPowerDown	4	static
+stm32f4xx_hal_pwr_ex.c:200:6:HAL_PWREx_DisableFlashPowerDown	4	static
+stm32f4xx_hal_pwr_ex.c:213:10:HAL_PWREx_GetVoltageRange	4	static
+stm32f4xx_hal_pwr_ex.c:290:19:HAL_PWREx_ControlVoltageScaling	32	static
+stm32f4xx_hal_pwr_ex.c:361:6:HAL_PWREx_EnableMainRegulatorLowVoltage	4	static
+stm32f4xx_hal_pwr_ex.c:372:6:HAL_PWREx_DisableMainRegulatorLowVoltage	4	static
+stm32f4xx_hal_pwr_ex.c:383:6:HAL_PWREx_EnableLowRegulatorLowVoltage	4	static
+stm32f4xx_hal_pwr_ex.c:394:6:HAL_PWREx_DisableLowRegulatorLowVoltage	4	static
