+++ host_comm_fsm/Debug/Drivers/STM32F4xx_HAL_Driver/Src/stm32f4xx_hal_rcc.su
@@ -1,0 +1,14 @@
+stm32f4xx_hal_rcc.c:202:26:HAL_RCC_DeInit	4	static
+stm32f4xx_hal_rcc.c:221:26:HAL_RCC_OscConfig	32	static
+stm32f4xx_hal_rcc.c:593:19:HAL_RCC_ClockConfig	24	static
+stm32f4xx_hal_rcc.c:777:6:HAL_RCC_MCOConfig	56	static
+stm32f4xx_hal_rcc.c:843:6:HAL_RCC_EnableCSS	4	static
+stm32f4xx_hal_rcc.c:852:6:HAL_RCC_DisableCSS	4	static
+stm32f4xx_hal_rcc.c:887:17:HAL_RCC_GetSysClockFreq	40	static
+stm32f4xx_hal_rcc.c:943:10:HAL_RCC_GetHCLKFreq	4	static
+stm32f4xx_hal_rcc.c:954:10:HAL_RCC_GetPCLK1Freq	8	static
+stm32f4xx_hal_rcc.c:966:10:HAL_RCC_GetPCLK2Freq	8	static
+stm32f4xx_hal_rcc.c:979:13:HAL_RCC_GetOscConfig	16	static
+stm32f4xx_hal_rcc.c:1058:6:HAL_RCC_GetClockConfig	16	static
+stm32f4xx_hal_rcc.c:1084:6:HAL_RCC_NMI_IRQHandler	8	static
+stm32f4xx_hal_rcc.c:1101:13:HAL_RCC_CSSCallback	4	static
