+++ host_comm_fsm/Debug/Drivers/STM32F4xx_HAL_Driver/Src/stm32f4xx_hal_rcc_ex.su
@@ -1,0 +1,7 @@
+stm32f4xx_hal_rcc_ex.c:2515:19:HAL_RCCEx_PeriphCLKConfig	32	static
+stm32f4xx_hal_rcc_ex.c:2643:6:HAL_RCCEx_GetPeriphCLKConfig	24	static
+stm32f4xx_hal_rcc_ex.c:2681:10:HAL_RCCEx_GetPeriphCLKFreq	32	static
+stm32f4xx_hal_rcc_ex.c:2768:6:HAL_RCCEx_SelectLSEMode	16	static
+stm32f4xx_hal_rcc_ex.c:2805:19:HAL_RCCEx_EnablePLLI2S	24	static
+stm32f4xx_hal_rcc_ex.c:2888:19:HAL_RCCEx_DisablePLLI2S	16	static
+stm32f4xx_hal_rcc_ex.c:3145:19:HAL_RCC_DeInit	16	static
