+++ host_comm_fsm/Debug/Drivers/STM32F4xx_HAL_Driver/Src/stm32f4xx_hal_uart.su
@@ -1,0 +1,61 @@
+stm32f4xx_hal_uart.c:356:19:HAL_UART_Init	16	static
+stm32f4xx_hal_uart.c:433:19:HAL_HalfDuplex_Init	16	static
+stm32f4xx_hal_uart.c:506:19:HAL_LIN_Init	16	static
+stm32f4xx_hal_uart.c:587:19:HAL_MultiProcessor_Init	24	static
+stm32f4xx_hal_uart.c:664:19:HAL_UART_DeInit	16	static
+stm32f4xx_hal_uart.c:709:13:HAL_UART_MspInit	16	static
+stm32f4xx_hal_uart.c:724:13:HAL_UART_MspDeInit	16	static
+stm32f4xx_hal_uart.c:1135:19:HAL_UART_Transmit	48	static
+stm32f4xx_hal_uart.c:1223:19:HAL_UART_Receive	48	static
+stm32f4xx_hal_uart.c:1314:19:HAL_UART_Transmit_IT	24	static
+stm32f4xx_hal_uart.c:1359:19:HAL_UART_Receive_IT	24	static
+stm32f4xx_hal_uart.c:1394:19:HAL_UART_Transmit_DMA	56	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:1462:19:HAL_UART_Receive_DMA	24	static
+stm32f4xx_hal_uart.c:1492:19:HAL_UART_DMAPause	120	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:1529:19:HAL_UART_DMAResume	120	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:1565:19:HAL_UART_DMAStop	72	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:1620:19:HAL_UARTEx_ReceiveToIdle	40	static
+stm32f4xx_hal_uart.c:1747:19:HAL_UARTEx_ReceiveToIdle_IT	56	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:1808:19:HAL_UARTEx_ReceiveToIdle_DMA	56	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:1865:19:HAL_UART_Abort	136	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:1954:19:HAL_UART_AbortTransmit	64	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:2005:19:HAL_UART_AbortReceive	112	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:2066:19:HAL_UART_Abort_IT	144	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:2201:19:HAL_UART_AbortTransmit_IT	64	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:2278:19:HAL_UART_AbortReceive_IT	112	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:2356:6:HAL_UART_IRQHandler	240	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:2588:13:HAL_UART_TxCpltCallback	16	static
+stm32f4xx_hal_uart.c:2603:13:HAL_UART_TxHalfCpltCallback	16	static
+stm32f4xx_hal_uart.c:2618:13:HAL_UART_RxCpltCallback	16	static
+stm32f4xx_hal_uart.c:2633:13:HAL_UART_RxHalfCpltCallback	16	static
+stm32f4xx_hal_uart.c:2648:13:HAL_UART_ErrorCallback	16	static
+stm32f4xx_hal_uart.c:2662:13:HAL_UART_AbortCpltCallback	16	static
+stm32f4xx_hal_uart.c:2677:13:HAL_UART_AbortTransmitCpltCallback	16	static
+stm32f4xx_hal_uart.c:2692:13:HAL_UART_AbortReceiveCpltCallback	16	static
+stm32f4xx_hal_uart.c:2709:13:HAL_UARTEx_RxEventCallback	16	static
+stm32f4xx_hal_uart.c:2749:19:HAL_LIN_SendBreak	40	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:2776:19:HAL_MultiProcessor_EnterMuteMode	40	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:2803:19:HAL_MultiProcessor_ExitMuteMode	40	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:2830:19:HAL_HalfDuplex_EnableTransmitter	24	static
+stm32f4xx_hal_uart.c:2865:19:HAL_HalfDuplex_EnableReceiver	24	static
+stm32f4xx_hal_uart.c:2922:23:HAL_UART_GetState	24	static
+stm32f4xx_hal_uart.c:2937:10:HAL_UART_GetError	16	static
+stm32f4xx_hal_uart.c:2982:13:UART_DMATransmitCplt	72	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:3017:13:UART_DMATxHalfCplt	24	static
+stm32f4xx_hal_uart.c:3036:13:UART_DMAReceiveCplt	120	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:3093:13:UART_DMARxHalfCplt	24	static
+stm32f4xx_hal_uart.c:3128:13:UART_DMAError	24	static
+stm32f4xx_hal_uart.c:3169:26:UART_WaitOnFlagUntilTimeout	72	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:3208:19:UART_Start_Receive_IT	24	static
+stm32f4xx_hal_uart.c:3243:19:UART_Start_Receive_DMA	104	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:3293:13:UART_EndTxTransfer	40	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:3307:13:UART_EndRxTransfer	88	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:3331:13:UART_DMAAbortOnError	24	static
+stm32f4xx_hal_uart.c:3355:13:UART_DMATxAbortCallback	24	static
+stm32f4xx_hal_uart.c:3401:13:UART_DMARxAbortCallback	24	static
+stm32f4xx_hal_uart.c:3447:13:UART_DMATxOnlyAbortCallback	24	static
+stm32f4xx_hal_uart.c:3475:13:UART_DMARxOnlyAbortCallback	24	static
+stm32f4xx_hal_uart.c:3501:26:UART_Transmit_IT	24	static
+stm32f4xx_hal_uart.c:3541:26:UART_EndTransmit_IT	16	static
+stm32f4xx_hal_uart.c:3566:26:UART_Receive_IT	56	static,ignoring_inline_asm
+stm32f4xx_hal_uart.c:3664:13:UART_SetConfig	160	static
